Orbit_API/requirements.txt
Django==2.2.9
djangorestframework==3.11.0
psycopg2==2.8.5
pytz==2019.3
sqlparse==0.3.1
